// src/components/panel/dto/customer-responsible-raw.dto.ts
export class CustomerResponsibleRawDto {
  customerid: string;
  customer_name: string;
  companyid: string;
  company_name: string;
  squaduserid: string;
  squaduser_name: string;
  reminderid: string;
  reminder_text: string;
  reminder_lastmodificationdate: Date;
  reminder_creationdate: Date;
}